Rest.Error

Description

Error types that can be returned by handlers, as well as some utilities for manipulating these errors.

Documentation

mapE :: (Applicative m, Monad m) => (e -> e') -> ErrorT e m a -> ErrorT e' m a infixl 8

orThrow :: MonadError e m => m (Maybe b) -> e -> m b

orThrowWith :: MonadError a m => m (Either e b) -> (e -> a) -> m b

eitherToStatus :: Either a b -> Status a b

domainReason :: (a -> Int) -> a -> Reason a

Wrap your custom error type in a Reason. The first argument is a function for converting your error to an HTTP status code.