Rest.Handler

Description

Handlers for endpoints in a Resource.

Single handlers.

mkHandler :: Monad m => Modifier h p i o e -> (Env h p i -> ErrorT (Reason e) m o) -> Handler m

Create a handler for a single resource. Takes the entire environmend as input.

mkInputHandler :: Monad m => Modifier () () i o e -> (i -> ErrorT (Reason e) m o) -> Handler m

Create a handler for a single resource. Takes only the body information as input.

mkConstHandler :: Monad m => Modifier () () () o e -> ErrorT (Reason e) m o -> Handler m

Create a handler for a single resource. Doesn't take any input.

mkIdHandler :: MonadReader id m => Modifier h p i o e -> (i -> id -> ErrorT (Reason e) m o) -> Handler m

Create a handler for a single resource. Take body information and the resource identifier as input. The monad m should be a Reader-like type containing the idenfier.

Listings.

mkListing :: Monad m => Modifier () () () o e -> (Range -> ErrorT (Reason e) m [o]) -> ListHandler m

Smart constructor for creating a list handler.

mkOrderedListing :: Monad m => Modifier () () () o e -> ((Range, Maybe String, Maybe String) -> ErrorT (Reason e) m [o]) -> ListHandler m

Create a list handler that accepts ordering information.

Parameter parsers for listings.

data Range

Data type for representing the requested range in list handlers.

Constructors

Range

Fields

offset :: Int

count :: Int

range :: Param Range

Dictionary for taking Range parameters. Allows two query parameters, offset and count. If not passed, the defaults are 0 and 100. The maximum range that can be passed is 1000.

orderedRange :: Param (Range, Maybe String, Maybe String)

Dictionary for taking ordering information. In addition to the parameters accepted by range, this accepts order and direction.

Generic handlers and core data types.

data Env h p i

An environment of inputs passed to a handler. Contains information from the headers, the parameters and the body input.

Constructors

Env

Fields

header :: h

param :: p

input :: i

data GenHandler m f where

A handler for some endpoint. The input and output types are specified by the dictionary, which can be created using the combinators from Rest.Dictionary.Combinators. The inputs (headers, parameters and body) are passed as an Env to the handler. This handler runs in monad m, combined with the ability to throw errors. The result is either the output value, or a list of them for list handlers. If the secure flag is set, this suggests to clients that the resource should only be served over https. It has no effect when running the API.

Constructors

GenHandler :: Dict h p i o e -> (Env h p i -> ErrorT (Reason e) m (Apply f o)) -> Bool -> GenHandler m f

Fields

dictionary :: Dict h p i o e

handler :: Env h p i -> ErrorT (Reason e) m (Apply f o)

secure :: Bool

mkGenHandler :: Monad m => Modifier h p i o e -> (Env h p i -> ErrorT (Reason e) m (Apply f o)) -> GenHandler m f

Construct a GenHandler using a Modifier instead of a Dict. The secure flag will be False.

type family Apply f a :: *

Apply a Functor f to a type a. In general will result in f a, except if f is Identity, in which case it will result in a. This prevents a lot of Identity wrapping/unwrapping.

type Handler m = GenHandler m Identity

A Handler returning a single item.

type ListHandler m = GenHandler m []

A Handler returning a list of items.

Convenience functions.

secureHandler :: Handler m -> Handler m

Set secure to True.